MAINT: Re-organize folder structure pt1
