+++ components/tabs_components.py
@@ -7,7 +7,7 @@ from components.dropdown_components import (
     custom_dropdow
 )
 from components.button_components import button, hourButton
-from styles.styles import button_style, button_dropdown_style
+from utils.styles import button_style, button_dropdown_style
 
 
 # Define a function to create the main tabs layout
